# Add static cache options