--- a/src/client/Microsoft.Identity.Client/Platforms/iOS/iOSPlatformProxy.cs
+++ b/src/client/Microsoft.Identity.Client/Platforms/iOS/iOSPlatformProxy.cs
@@ -91,7 +91,7 @@ namespace Microsoft.Identity.Client.Platforms.iOS
             return new iOSLegacyCachePersistence(Logger);
         }
 
-        public override ITokenCacheAccessor CreateTokenCacheAccessor(bool isApplicationTokenCache = false)
+        public override ITokenCacheAccessor CreateTokenCacheAccessor(InternalMemoryTokenCacheOptions tokenCacheAccessorOptions, bool isApplicationTokenCache = false)
         {
             return new iOSTokenCacheAccessor();
         }