--- a/src/client/Microsoft.Identity.Client/TokenCache.Notifications.cs
+++ b/src/client/Microsoft.Identity.Client/TokenCache.Notifications.cs
@@ -38,7 +38,7 @@ namespace Microsoft.Identity.Client
         bool ITokenCacheInternal.IsExternalSerializationConfiguredByUser()
         {
             return !this.UsesDefaultSerialization &&
-                (this as ITokenCacheInternal).IsAppSubscribedToSerializationEvents();                
+                (this as ITokenCacheInternal).IsAppSubscribedToSerializationEvents();
         }
 
         async Task ITokenCacheInternal.OnAfterAccessAsync(TokenCacheNotificationArgs args)
@@ -89,8 +89,7 @@ namespace Microsoft.Identity.Client
 #endif
         public void SetBeforeAccess(TokenCacheCallback beforeAccess)
         {
-            GuardOnMobilePlatforms();
-
+            Validate();
             ResetDefaultDelegates();
             BeforeAccess = beforeAccess;
         }
@@ -109,7 +108,7 @@ namespace Microsoft.Identity.Client
 #endif
         public void SetAfterAccess(TokenCacheCallback afterAccess)
         {
-            GuardOnMobilePlatforms();
+            Validate();
             ResetDefaultDelegates();
             AfterAccess = afterAccess;
         }
@@ -125,7 +124,7 @@ namespace Microsoft.Identity.Client
 #endif
         public void SetBeforeWrite(TokenCacheCallback beforeWrite)
         {
-            GuardOnMobilePlatforms();
+            Validate();
             ResetDefaultDelegates();
             BeforeWrite = beforeWrite;
         }
@@ -139,7 +138,7 @@ namespace Microsoft.Identity.Client
 #endif
         public void SetBeforeAccessAsync(Func<TokenCacheNotificationArgs, Task> beforeAccess)
         {
-            GuardOnMobilePlatforms();
+            Validate();
             ResetDefaultDelegates();
             AsyncBeforeAccess = beforeAccess;
         }
@@ -153,7 +152,7 @@ namespace Microsoft.Identity.Client
 #endif
         public void SetAfterAccessAsync(Func<TokenCacheNotificationArgs, Task> afterAccess)
         {
-            GuardOnMobilePlatforms();
+            Validate();
             ResetDefaultDelegates();
             AsyncAfterAccess = afterAccess;
         }
@@ -167,13 +166,20 @@ namespace Microsoft.Identity.Client
 #endif
         public void SetBeforeWriteAsync(Func<TokenCacheNotificationArgs, Task> beforeWrite)
         {
-            GuardOnMobilePlatforms();
+            Validate();
             ResetDefaultDelegates();
             AsyncBeforeWrite = beforeWrite;
         }
 
-        private static void GuardOnMobilePlatforms()
+        private void Validate()
         {
+            if (ServiceBundle.Config.AccessorOptions != null)
+            {
+                throw new MsalClientException(
+                    MsalError.StaticCacheWithExternalSerialization,
+                    MsalErrorMessage.StaticCacheWithExternalSerialization);
+            }
+
 #if !SUPPORTS_CUSTOM_CACHE
         throw new PlatformNotSupportedException("You should not use these TokenCache methods on mobile platforms. " +
             "They are meant to allow applications to define their own storage strategy on .net desktop and non-mobile platforms such as .net core. " +