… tests/Microsoft.Identity.Test.Common/Core/Helpers/UserAccessorWithPartitionAsserts.cs
… tests/Microsoft.Identity.Test.Unit/CacheTests/CacheSerializationTests.cs
… tests/Microsoft.Identity.Test.Unit/CacheTests/InMemoryPartitionedTokenCacheAccessorTests.cs
… tests/Microsoft.Identity.Test.Unit/CacheTests/InternalCacheOptionsTests.cs
… tests/Microsoft.Identity.Test.Unit/CoreTests/PlatformProxyFactoryTests.cs
… tests/devapps/WebApi/Controllers/FlatCacheController.cs
… tests/devapps/WebApi/Controllers/OboController.cs
… tests/devapps/WebApi/Controllers/PrototypeController.cs
… tests/devapps/WebApi/Controllers/WilsonLruCacheController.cs
… tests/devapps/WebApi/Misc/AbstractPartitionedCacheSerializer.cs
… tests/devapps/WebApi/Misc/InMemoryPartitionedCacheSerializer.cs