Fixes #2849
<!-- Add the issue number above. If this PR only partially fixes the issue, remove the Fixes word above so that the issue is not automatically closed. -->

**Changes proposed in this request**
<!-- Concisely list the changes. -->
<!-- If helpful, describe how and why the bug was fixed. -->
<!-- If needed, describe how to review (ex. which files have the primary changes, which are nit changes, etc.) -->

**Testing**
ut + perf

**Performance impact** No difference


### With static cache

|                                    Method | Tenants | TokensPerTenant |     Mean |     Error |    StdDev |    Median |       Min |      Max | Allocated |
|------------------------------------------ |-------- |---------------- |---------:|----------:|----------:|----------:|----------:|---------:|----------:|
| AcquireTokenForClient_Singleton_TestAsync |    1000 |            1000 | 1.029 ms | 0.0878 ms | 0.2546 ms | 0.9091 ms | 0.7325 ms | 1.746 ms |    185 KB |

### Baseline

|                          Method | Tenants | TokensPerTenant |     Mean |    Error |   StdDev |   Median |      Min |        Max | Allocated |
|-------------------------------- |-------- |---------------- |---------:|---------:|---------:|---------:|---------:|-----------:|----------:|
| AcquireTokenForClient_TestAsync |    1000 |            1000 | 956.9 us | 76.79 us | 215.3 us | 865.7 us | 661.6 us | 1,626.7 us |    185 KB |