@trwalke 2021-09-13
> <Project Sdk="MSBuild.Sdk.Extras">

is there something different in this file? i dont see any changes

---
Refers to: src/client/Microsoft.Identity.Client/Microsoft.Identity.Client.csproj:1 in bfc25ae. [](commit_id = bfc25aee18151a3b8a62f43c83fa76718c68797b, deletion_comment = False)